[FEATURE]: decomposed reducer
