+++ src/reducer/namespace.js
@@ -1,0 +1,4 @@
+export default {
+  DATA: `DATA`,
+  GAME: `GAME`
+};
